requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
ultralytics>=8.3.0
opencv-python>=4.9.0
pillow>=10.2.0
numpy>=1.26.0

